Move SeamApiException to exceptions.py and update imports
