+++ seam/__init__.py
@@ -2,11 +2,11 @@
 # type: ignore
 
 from .seam import Seam
-from .models import SeamApiException
 from .seam_multi_workspace import SeamMultiWorkspace
 from .options import SeamHttpInvalidOptionsError
 from .auth import SeamHttpInvalidTokenError
 from .exceptions import (
+    SeamApiException,
     SeamActionAttemptError,
     SeamActionAttemptFailedError,
     SeamActionAttemptTimeoutError,
