+++ seam/exceptions.py
@@ -1,4 +1,24 @@
+from niquests import Response
 from .routes.models import ActionAttempt
+
+
+# HTTP
+class SeamApiException(Exception):
+    def __init__(
+        self,
+        response: Response,
+    ):
+        self.status_code = response.status_code
+        self.request_id = response.headers.get("seam-request-id", None)
+
+        self.metadata = None
+        if "application/json" in response.headers["content-type"]:
+            parsed_response = response.json()
+            self.metadata = parsed_response.get("error", None)
+
+        super().__init__(
+            f"SeamApiException: status={self.status_code}, request_id={self.request_id}, metadata={self.metadata}"
+        )
 
 
 # Action Attempt
