+++ seam/models.py
@@ -1,27 +1,8 @@
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict, List, Optional, Union
 from typing_extensions import Self
 import abc
-from niquests import Response
 
 from .routes.models import AbstractRoutes, Workspace
-
-
-class SeamApiException(Exception):
-    def __init__(
-        self,
-        response: Response,
-    ):
-        self.status_code = response.status_code
-        self.request_id = response.headers.get("seam-request-id", None)
-
-        self.metadata = None
-        if "application/json" in response.headers["content-type"]:
-            parsed_response = response.json()
-            self.metadata = parsed_response.get("error", None)
-
-        super().__init__(
-            f"SeamApiException: status={self.status_code}, request_id={self.request_id}, metadata={self.metadata}"
-        )
 
 
 class AbstractRequestMixin(abc.ABC):
