+++ test/access_codes/test_access_codes.py
@@ -1,5 +1,5 @@
 from seam import Seam
-from seam.types import SeamApiException
+from seam.exceptions import SeamApiException
 import pytest
